+++ core/src/main/groovy/com/ainrif/apiator/doclet/ApiatorDoclet.groovy
@@ -16,22 +16,99 @@
 
 package com.ainrif.apiator.doclet
 
-import com.sun.javadoc.DocErrorReporter
-import com.sun.javadoc.LanguageVersion
-import com.sun.javadoc.RootDoc
-import com.sun.tools.javadoc.Main
 import groovy.json.JsonBuilder
+import groovy.transform.CompileStatic
 import groovy.transform.Immutable
+import groovy.util.logging.Slf4j
+import jdk.javadoc.doclet.Doclet
+import jdk.javadoc.doclet.DocletEnvironment
+import jdk.javadoc.doclet.Reporter
 
 import javax.annotation.Nullable
+import javax.lang.model.SourceVersion
+import javax.tools.Diagnostic
+import javax.tools.ToolProvider
 
 import static java.io.File.createTempFile
 
-/**
- * @see com.sun.javadoc.Doclet
- */
-class ApiatorDoclet {
+@Slf4j
+@CompileStatic
+class ApiatorDoclet implements Doclet {
     public static final String OF_PARAM = '--output-file'
+
+    private Reporter reporter
+    private String outputFilePath
+
+    @Override
+    void init(Locale locale, Reporter reporter) {
+        this.reporter = reporter
+    }
+
+    @Override
+    String getName() {
+        return 'Apiator'
+    }
+
+    @Override
+    SourceVersion getSupportedSourceVersion() {
+        return SourceVersion.RELEASE_11
+    }
+
+    @Override
+    boolean run(DocletEnvironment environment) {
+        def docInfo = new JavaDocInfoBuilder(environment).content
+        new File(outputFilePath).write(new JsonBuilder(docInfo).toString())
+
+        return true
+    }
+
+    @Override
+    Set<? extends Option> getSupportedOptions() {
+        def ofOption = new Option() {
+            @Override
+            int getArgumentCount() {
+                return 1
+            }
+
+            @Override
+            String getDescription() {
+                return 'Output file path'
+            }
+
+            @Override
+            Option.Kind getKind() {
+                return Option.Kind.STANDARD
+            }
+
+            @Override
+            List<String> getNames() {
+                return List.of(OF_PARAM)
+            }
+
+            @Override
+            String getParameters() {
+                return 'file'
+            }
+
+            @Override
+            boolean process(String option, List<String> arguments) {
+                if (arguments) {
+                    outputFilePath = arguments[0]
+
+                    if (!new File(outputFilePath).canWrite()) {
+                        reporter.print(Diagnostic.Kind.ERROR, "Cannot write to file: ${outputFilePath}")
+                        return false
+                    }
+                } else {
+                    reporter.print(Diagnostic.Kind.ERROR, "File path is absent")
+                }
+
+                return true
+            }
+        }
+
+        return Set.of(ofOption)
+    }
 
     /**
      * @param sourcePath
@@ -42,10 +119,9 @@ class ApiatorDoclet {
      */
     static Result runDoclet(String sourcePath,
                             @Nullable String docletClassPath,
-                            @Nullable String basePackage,
+                            String basePackage,
                             @Nullable String outputFile) {
         outputFile = outputFile ?: createTempFile('apiator', 'doclet').with { it.deleteOnExit(); it }.absolutePath
-        basePackage = basePackage ?: '.'
 
         def javaDocArgs = ['-sourcepath', sourcePath,
                            '-doclet', ApiatorDoclet.class.name,
@@ -58,55 +134,27 @@ class ApiatorDoclet {
             javaDocArgs.add(docletClassPath)
         }
 
-        return new Result(Main.execute(javaDocArgs as String[]), outputFile)
-    }
+        try {
+            boolean success = ToolProvider
+                    .getSystemDocumentationTool()
+                    .getTask(null, null, null, ApiatorDoclet, javaDocArgs, null)
+                    .call()
 
-    static boolean start(RootDoc root) {
-        def docInfo = new JavaDocInfoBuilder(root.classes()).content
-
-        def filePath = getOptionValue(root.options(), OF_PARAM)[1]
-        new File(filePath).write(new JsonBuilder(docInfo).toString())
-
-        return true
-    }
-
-    static int optionLength(String option) {
-        switch (option) {
-            case OF_PARAM:
-                return 2
-            default:
-                return 0
-        }
-    }
-
-    static boolean validOptions(String[][] options, DocErrorReporter reporter) {
-        String[] ofArray = getOptionValue(options, OF_PARAM)
-        if (!(ofArray && ofArray[1])) {
-            reporter.printError("Required parameter is absent: ${OF_PARAM} <file>")
-            return false
-        } else {
-            def of = ofArray[1]
-            if (!new File(of).canWrite()) {
-                reporter.printError("Cannot write to file: ${of}")
-                return false
+            if (!success) {
+                log.warn('Result of doclet was not successful and there are no exceptions')
             }
+
+            new Result(success, outputFile)
+        } catch (Throwable e) {
+            log.error('Error during doclet executing.', e)
+
+            new Result(false, outputFile)
         }
-
-        return true
-    }
-
-    static LanguageVersion languageVersion() {
-        return LanguageVersion.JAVA_1_5
     }
 
     @Immutable
     static class Result {
-        int code
+        boolean success
         String outputFile
-    }
-
-    @Nullable
-    private static String[] getOptionValue(String[][] options, String name) {
-        return options.find { it[0] == name }
     }
 }