+++ src/test/groovy/com/ainrif/apiator/test/smoke/CoreJsonRendererSmokeSpec.groovy
@@ -46,7 +46,8 @@ class CoreJsonRendererSmokeSpec extends Specification {
         return new ApiatorConfig(
                 provider: new JaxRsProvider(),
                 docletConfig: new DocletConfig(
-                        sourcePath: buildSourcePath()
+                        sourcePath: buildSourcePath(),
+                        includeBasePackage: 'com.ainrif'
                 ),
                 renderer: new CoreJsonRenderer({
                     plugins << new JaxRsCompositePlugin()